LayoutTests/accessibility/text-marker/text-marker-previous-next.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<style>
.userselect { user-select: none; -webkit-user-select: none; }
</style>
<body id="body">

<div id="text" tabindex="0">text</div>
<br>
text1

<div id="text2">
c <img src="#" aria-label="blah" style="background-color: #aaaaaa; width: 100px; height: 100px;"> d
</div>

<input type="password" id="psw">

<div class="userselect" id="text3">can't select</div>

<div id="text4">
abc
de
f
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests the next/previous text marker functions are implemented correctly.");
    
    if (window.accessibilityController) {
        
        var text = accessibilityController.accessibleElementById("text");
        // Get the actual text node.
        text = text.childAtIndex(0);
        
        // Check that we can get the start/end marker for this range.
        var textMarkerRange = text.textMarkerRangeForElement(text);
        shouldBe("text.textMarkerRangeLength(textMarkerRange)", "4");
        
        var startMarker = text.startTextMarkerForTextMarkerRange(textMarkerRange);
        var endMarker = text.endTextMarkerForTextMarkerRange(textMarkerRange);
        shouldBeTrue("text.accessibilityElementForTextMarker(startMarker).isEqual(text)");
        shouldBeTrue("text.accessibilityElementForTextMarker(endMarker).isEqual(text)");
        
        // Check next text marker. (Advance 5 characters, it will land at <br>.)
        var currentMarker = startMarker;
        var previousMarker, markerRange;
        for (var i = 0; i < 5; i++) {
            previousMarker = currentMarker;
            currentMarker = text.nextTextMarker(currentMarker);
        }
        markerRange = text.textMarkerRangeForMarkers(previousMarker, currentMarker);
        var newline = '\n';
        shouldBe("text.stringForTextMarkerRange(markerRange)", "newline");
        
        // Advance one more character, it will lande at "t" in "text1".
        previousMarker = currentMarker;
        currentMarker = text.nextTextMarker(currentMarker);
        markerRange = text.textMarkerRangeForMarkers(previousMarker, currentMarker);
        shouldBe("text.stringForTextMarkerRange(markerRange)", "'t'");
        
        // Check previous text marker. (Traverse backwards one character, it will land at <br>.)
        previousMarker = text.previousTextMarker(previousMarker);
        currentMarker = text.previousTextMarker(currentMarker);
        markerRange = text.textMarkerRangeForMarkers(previousMarker, currentMarker);
        shouldBe("text.stringForTextMarkerRange(markerRange)", "newline");
        
        // Traverse backwards one more character, it will land at the last character of "text".
        previousMarker = text.previousTextMarker(previousMarker);
        currentMarker = text.previousTextMarker(currentMarker);
        markerRange = text.textMarkerRangeForMarkers(previousMarker, currentMarker);
        shouldBe("text.stringForTextMarkerRange(markerRange)", "'t'");
        
        // Check the case with replaced node
        var text2 = accessibilityController.accessibleElementById("text2");
        var textMarkerRange2 = text2.textMarkerRangeForElement(text2);
        shouldBe("text2.textMarkerRangeLength(textMarkerRange2)", "5");
        var str = text2.stringForTextMarkerRange(textMarkerRange2).replace(String.fromCharCode(65532), "[ATTACHMENT]");
        debug("Object string for range: " + str);
        
        currentMarker = text2.startTextMarkerForTextMarkerRange(textMarkerRange2);
        // Advance 5 characters, it will land at "d".
        for (var i = 0; i < 5; i++) {
            previousMarker = currentMarker;
            currentMarker = text2.nextTextMarker(currentMarker);
        }
        markerRange = text2.textMarkerRangeForMarkers(previousMarker, currentMarker);
        shouldBe("text2.stringForTextMarkerRange(markerRange)", "'d'");
        
        // Traverse backwards 5 characters, it will land at the last character of "text1".
        for (var i = 0; i < 5; i++) {
            previousMarker = text2.previousTextMarker(previousMarker);
            currentMarker = text2.previousTextMarker(currentMarker);
        }
        markerRange = text2.textMarkerRangeForMarkers(previousMarker, currentMarker);
        shouldBe("text2.stringForTextMarkerRange(markerRange)", "'1'");
        
        // Check the case with user-select:none, nextTextMarker/previousTextMarker should still work.
        var text3 = accessibilityController.accessibleElementById("text3");
        text3 = text3.childAtIndex(0);
        // Advance to land at user-select:none node.
        var marker1, marker2;
        for (var i = 0; i < 17; i++) {
            currentMarker = text3.nextTextMarker(currentMarker);
            // i == 13, it should land on "e", and i == 16, it should land on "t"
            if (i == 13) {
                marker1 = currentMarker;
            }
        }
        marker2 = currentMarker;
        markerRange = text3.textMarkerRangeForMarkers(marker1, marker2);
        shouldBe("text3.stringForTextMarkerRange(markerRange)", "'ect'");
        // Iterate backwards the second marker for 6 characters, the range should be "sel"
        for (var i = 0; i < 6; i++) {
            currentMarker = text3.previousTextMarker(currentMarker);
        }
        marker2 = currentMarker;
        markerRange = text3.textMarkerRangeForMarkers(marker1, marker2);
        shouldBe("text3.stringForTextMarkerRange(markerRange)", "'sel'");

        // Check the case with password field.
        var psw = accessibilityController.accessibleElementById("psw");
        var textMarkerRange3 = psw.textMarkerRangeForElement(psw);
        var start = psw.startTextMarkerForTextMarkerRange(textMarkerRange3);
        shouldBeTrue("!psw.accessibilityElementForTextMarker(start)");
        
        // Check next/previous text marker call will skip password field
        // We start from text2 and advance 6 characters, it should skip the password field and land on text3.
        currentMarker = text2.startTextMarkerForTextMarkerRange(textMarkerRange2);
        for (var i = 0; i < 6; i++) {
            currentMarker = text2.nextTextMarker(currentMarker);
        }
        shouldBeTrue("text2.accessibilityElementForTextMarker(currentMarker).isEqual(text3)");
        // Check previous text marker, it should land on " d" node.
        currentMarker = text2.previousTextMarker(currentMarker);
        shouldBeTrue("text2.accessibilityElementForTextMarker(currentMarker).isEqual(text2.childAtIndex(2))");
        
        
        // Make sure that text node with line breaks, we can go through it with next/previous call.
        text = accessibilityController.accessibleElementById("text4");
        textMarkerRange = text.textMarkerRangeForElement(text);
        startMarker = text.startTextMarkerForTextMarkerRange(textMarkerRange);
        currentMarker = startMarker;
        for (var i = 0; i < 8; i++) {
            previousMarker = currentMarker;
            currentMarker = text.nextTextMarker(currentMarker);
        }
        markerRange = text.textMarkerRangeForMarkers(previousMarker, currentMarker)
        shouldBe("text.stringForTextMarkerRange(markerRange)", "'f'");
        
        endMarker = text.endTextMarkerForTextMarkerRange(textMarkerRange);
        currentMarker = endMarker;
        for (var i = 0; i < 7; i++) {
            currentMarker = text.previousTextMarker(currentMarker);
        }
        markerRange = text.textMarkerRangeForMarkers(startMarker, currentMarker)
        shouldBe("text.stringForTextMarkerRange(markerRange)", "'a'");
        
    }

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
